+++ app/src/main/res/layout/fragment_group.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context="tech.jianka.fragment.TabsFragment">
+    tools:context="tech.jianka.fragment.GroupFragment">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/group_recycler_view"
