+++ app/src/main/res/layout/fragment_recent.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context="tech.jianka.fragment.TabsFragment">
+    tools:context="tech.jianka.fragment.RecentFragment">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/recent_recycler_view"
